Minor Changes in Code Relative to Proper Formatting

+++ src/main/webapp/WEB-INF/views/HomePage.jsp
@@ -109,7 +109,7 @@ table th {
         <td><c:out value="${tempCustomer.emailId}" /></td>
         <td>
           <a href="/CustomerManagement/updateRecord?id=${tempCustomer.id}"><button>UPDATE</button></a>
-          <a href="/CustomerManagement/deleteRecord?id=${tempCustomer.id}"><button>DELETE</button></a>
+          <a href="/CustomerManagement/deleteRecord?id=${tempCustomer.id}" onclick="if(!(confirm('Are you sure you want to delete the record ?')))return false"><button>DELETE</button></a>
         </td>
       </tr>
    </c:forEach>
